Fixed linting errors and Android gestures, ready for PR
