+++ mobile/.eslintrc.json
@@ -14,6 +14,7 @@
   },
   "plugins": ["react", "@typescript-eslint"],
   "rules": {
-    "react/style-prop-object": "off"
+    "react/style-prop-object": "off",
+    "react/no-array-index-key": "off"
   }
 }
