+++ mobile/babel.config.js
@@ -2,6 +2,6 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: ["babel-preset-expo"],
-    plugins: ['react-native-reanimated/plugin'],
+    plugins: ["react-native-reanimated/plugin"],
   };
 };
