--- mobile/index.ts
@@ -1,4 +1,0 @@
-export { default } from "./Chess";
-import { PIECES } from "./Piece";
-
-export const assets = Object.values(PIECES);
