+++ mobile/package.json
@@ -21,6 +21,7 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-gesture-handler": "^1.10.3",
+    "react-native-navbar": "^2.1.0",
     "react-native-reanimated": "^2.3.1",
     "react-native-redash": "^16.2.3",
     "react-native-web": "0.17.1"
