+++ mobile/src/screens/components/Background.tsx
@@ -23,7 +23,7 @@ interface SquareProps extends RowProps {
   col: number;
 }
 
-const Square = ({ white, row, col }: SquareProps) => {
+function Square({ white, row, col }: SquareProps) {
   const backgroundColor = white ? WHITE : BLACK;
   const color = white ? BLACK : WHITE;
   const textStyle = { fontWeight: "500" as const, color };
@@ -36,19 +36,15 @@ const Square = ({ white, row, col }: SquareProps) => {
         justifyContent: "space-between",
       }}
     >
-      <Text style={[textStyle, { opacity: col === 0 ? 1 : 0 }]}>
-        {"" + (8 - row)}
-      </Text>
+      <Text style={[textStyle, { opacity: col === 0 ? 1 : 0 }]}>{"" + (8 - row)}</Text>
       {row === 7 && (
-        <Text style={[textStyle, { alignSelf: "flex-end" }]}>
-          {String.fromCharCode(97 + col)}
-        </Text>
+        <Text style={[textStyle, { alignSelf: "flex-end" }]}>{String.fromCharCode(97 + col)}</Text>
       )}
     </View>
   );
-};
+}
 
-const Row = ({ white, row }: RowProps) => {
+function Row({ white, row }: RowProps) {
   const offset = white ? 0 : 1;
   return (
     <View style={styles.container}>
@@ -57,9 +53,9 @@ const Row = ({ white, row }: RowProps) => {
       ))}
     </View>
   );
-};
+}
 
-const Background = () => {
+function Background() {
   return (
     <View style={{ flex: 1 }}>
       {new Array(8).fill(0).map((_, i) => (
@@ -67,6 +63,6 @@ const Background = () => {
       ))}
     </View>
   );
-};
+}
 
 export default Background;
