+++ mobile/src/screens/components/Piece.tsx
@@ -11,7 +11,21 @@ import Animated, {
 } from "react-native-reanimated";
 import { Vector } from "react-native-redash";
 
-import { toTranslation, SIZE, toPosition } from "./Notation";
+// import black piece image assets
+import bbPiece from "../../../assets/bb.png";
+import bkPiece from "../../../assets/bk.png";
+import bnPiece from "../../../assets/bn.png";
+import bpPiece from "../../../assets/bp.png";
+import bqPiece from "../../../assets/bq.png";
+import brPiece from "../../../assets/br.png";
+import wbPiece from "../../../assets/wb.png";
+import wkPiece from "../../../assets/wk.png";
+import wnPiece from "../../../assets/wn.png";
+import wpPiece from "../../../assets/wp.png";
+import wqPiece from "../../../assets/wq.png";
+import wrPiece from "../../../assets/wr.png";
+
+import { toTranslation, SIZE, toPosition } from "./Notation.ts";
 
 const styles = StyleSheet.create({
   piece: {
@@ -21,32 +35,32 @@ const styles = StyleSheet.create({
 });
 type Player = "b" | "w";
 type Type = "q" | "r" | "n" | "b" | "k" | "p";
-type Piece = `${Player}${Type}`;
-type Pieces = Record<Piece, ReturnType<typeof require>>;
+type PlayerPiece = `${Player}${Type}`;
+type Pieces = Record<PlayerPiece, ReturnType<typeof require>>;
 export const PIECES: Pieces = {
-  br: require("./assets/br.png"),
-  bp: require("./assets/bp.png"),
-  bn: require("./assets/bn.png"),
-  bb: require("./assets/bb.png"),
-  bq: require("./assets/bq.png"),
-  bk: require("./assets/bk.png"),
-  wr: require("./assets/wr.png"),
-  wn: require("./assets/wn.png"),
-  wb: require("./assets/wb.png"),
-  wq: require("./assets/wq.png"),
-  wk: require("./assets/wk.png"),
-  wp: require("./assets/wp.png"),
+  br: brPiece,
+  bp: bpPiece,
+  bn: bnPiece,
+  bb: bbPiece,
+  bq: bqPiece,
+  bk: bkPiece,
+  wr: wrPiece,
+  wn: wnPiece,
+  wb: wbPiece,
+  wq: wqPiece,
+  wk: wkPiece,
+  wp: wpPiece,
 };
 
 interface PieceProps {
-  id: Piece;
+  id: PlayerPiece;
   startPosition: Vector;
   chess: Chess;
   onTurn: () => void;
   enabled: boolean;
 }
 
-const Piece = ({ id, startPosition, chess, onTurn, enabled }: PieceProps) => {
+function Piece({ id, startPosition, chess, onTurn, enabled }: PieceProps) {
   const isGestureActive = useSharedValue(false);
   const offsetX = useSharedValue(0);
   const offsetY = useSharedValue(0);
@@ -60,11 +74,9 @@ const Piece = ({ id, startPosition, chess, onTurn, enabled }: PieceProps) => {
       const from = toPosition({ x: offsetX.value, y: offsetY.value });
       const move = moves.find((m) => m.from === from && m.to === to);
       const { x, y } = toTranslation(move ? move.to : from);
-      translateX.value = withTiming(
-        x,
-        {},
-        () => (offsetX.value = translateX.value)
-      );
+      translateX.value = withTiming(x, {}, () => {
+        offsetX.value = translateX.value;
+      });
       translateY.value = withTiming(y, {}, () => {
         offsetY.value = translateY.value;
         isGestureActive.value = false;
@@ -87,31 +99,22 @@ const Piece = ({ id, startPosition, chess, onTurn, enabled }: PieceProps) => {
       translateY.value = offsetY.value + translationY;
     },
     onEnd: () => {
-      runOnJS(movePiece)(
-        toPosition({ x: translateX.value, y: translateY.value })
-      );
+      runOnJS(movePiece)(toPosition({ x: translateX.value, y: translateY.value }));
     },
   });
   const style = useAnimatedStyle(() => ({
     position: "absolute",
     zIndex: isGestureActive.value ? 100 : 10,
-    transform: [
-      { translateX: translateX.value },
-      { translateY: translateY.value },
-    ],
+    transform: [{ translateX: translateX.value }, { translateY: translateY.value }],
   }));
-  const original = useAnimatedStyle(() => {
-    return {
-      position: "absolute",
-      width: SIZE,
-      height: SIZE,
-      zIndex: 0,
-      backgroundColor: isGestureActive.value
-        ? "rgba(255, 255, 0, 0.5)"
-        : "transparent",
-      transform: [{ translateX: offsetX.value }, { translateY: offsetY.value }],
-    };
-  });
+  const original = useAnimatedStyle(() => ({
+    position: "absolute",
+    width: SIZE,
+    height: SIZE,
+    zIndex: 0,
+    backgroundColor: isGestureActive.value ? "rgba(255, 255, 0, 0.5)" : "transparent",
+    transform: [{ translateX: offsetX.value }, { translateY: offsetY.value }],
+  }));
   const underlay = useAnimatedStyle(() => {
     const position = toPosition({ x: translateX.value, y: translateY.value });
     const translation = toTranslation(position);
@@ -120,9 +123,7 @@ const Piece = ({ id, startPosition, chess, onTurn, enabled }: PieceProps) => {
       width: SIZE,
       height: SIZE,
       zIndex: 0,
-      backgroundColor: isGestureActive.value
-        ? "rgba(255, 255, 0, 0.5)"
-        : "transparent",
+      backgroundColor: isGestureActive.value ? "rgba(255, 255, 0, 0.5)" : "transparent",
       transform: [{ translateX: translation.x }, { translateY: translation.y }],
     };
   });
@@ -137,6 +138,6 @@ const Piece = ({ id, startPosition, chess, onTurn, enabled }: PieceProps) => {
       </PanGestureHandler>
     </>
   );
-};
+}
 
 export default Piece;
